migrate showLanguageSwith to his own Combo Interface/Trait. Implemented in FormSidebarWidget and QuickFormWidget
