+++ src/Charcoal/Admin/Ui/HasLanguageSwitcherInterface.php
@@ -1,0 +1,24 @@
+<?php
+
+namespace Charcoal\Admin\Ui;
+
+/**
+ *
+ */
+interface HasLanguageSwitcherInterface
+{
+    /**
+     * Whether to display the language switcher.
+     *
+     * @return boolean
+     */
+    public function showLanguageSwitch();
+
+    /**
+     * Retrieve the available languages, formatted for the sidebar language-switcher.
+     *
+     * @see    FormGroupWidget::languages()
+     * @return array
+     */
+    public function languages();
+}
