+++ src/Charcoal/Admin/Widget/FormSidebarWidget.php
@@ -11,6 +11,8 @@ use Pimple\Container;
 
 // From 'charcoal-ui'
 use Charcoal\Ui\Form\FormInterface;
+use Charcoal\Admin\Ui\HasLanguageSwitcherInterface;
+use Charcoal\Admin\Ui\HasLanguageSwitcherTrait;
 
 // From 'charcoal-translator'
 use Charcoal\Translator\Translation;
@@ -24,9 +26,11 @@ use Charcoal\Admin\Ui\FormSidebarInterface;
  * Form Sidebar Widget
  */
 class FormSidebarWidget extends AdminWidget implements
-    FormSidebarInterface
+    FormSidebarInterface,
+    HasLanguageSwitcherInterface
 {
     use ActionContainerTrait;
+    use HasLanguageSwitcherTrait;
 
     /**
      * Default sorting priority for an action.
@@ -118,13 +122,6 @@ class FormSidebarWidget extends AdminWidget implements
      * @var boolean
      */
     protected $showSidebarActions;
-
-    /**
-     * Whether to display the language switcher.
-     *
-     * @var boolean
-     */
-    protected $showLanguageSwitch;
 
     /**
      * Whether the object is viewable.
@@ -796,82 +793,17 @@ class FormSidebarWidget extends AdminWidget implements
     }
 
     /**
-     * @see    FormPropertyWidget::showActiveLanguage()
+     * @see    HasLanguageSwitcherTrait::showLanguageSwitch()
      * @return boolean
      */
-    public function showLanguageSwitch()
+    protected function resolveShowLanguageSwitch()
     {
-        if ($this->showLanguageSwitch === null) {
-            $form = $this->form();
-            if ($form) {
-                $this->showLanguageSwitch = $form->hasL10nFormProperties();
-                return $this->showLanguageSwitch;
-            }
-
-            $this->showLanguageSwitch = false;
+        $form = $this->form();
+        if ($form) {
+            return $form->hasL10nFormProperties();
         }
 
-        return $this->showLanguageSwitch;
-    }
-
-    /**
-     * Retrieve the available languages, formatted for the sidebar language-switcher.
-     *
-     * @see    FormGroupWidget::languages()
-     * @return array
-     */
-    public function languages()
-    {
-        $currentLocale = $this->translator()->getLocale();
-        $locales = $this->translator()->locales();
-        $languages = [];
-
-        uasort($locales, [ $this, 'sortLanguagesByPriority' ]);
-
-        foreach ($locales as $locale => $localeStruct) {
-            /**
-             * @see \Charcoal\Admin\Widget\FormGroupWidget::languages()
-             * @see \Charcoal\Property\LangProperty::localeChoices()
-             */
-            if (isset($localeStruct['name'])) {
-                $label = $this->translator()->translation($localeStruct['name']);
-            } else {
-                $trans = 'locale.'.$locale;
-                if ($trans === $this->translator()->trans($trans)) {
-                    $label = strtoupper($locale);
-                } else {
-                    $label = $this->translator()->translation($trans);
-                }
-            }
-
-            $isCurrent = ($locale === $currentLocale);
-            $languages[] = [
-                'cssClasses' => ($isCurrent) ? 'btn-primary' : 'btn-outline-primary',
-                'ident'      => $locale,
-                'name'       => $label,
-                'current'    => $isCurrent
-            ];
-        }
-
-        return $languages;
-    }
-
-    /**
-     * To be called with {@see uasort()}.
-     *
-     * @param  array $a Sortable action A.
-     * @param  array $b Sortable action B.
-     * @return integer
-     */
-    protected function sortLanguagesByPriority(array $a, array $b)
-    {
-        $a = isset($a['priority']) ? $a['priority'] : 0;
-        $b = isset($b['priority']) ? $b['priority'] : 0;
-
-        if ($a === $b) {
-            return 0;
-        }
-        return ($a < $b) ? (-1) : 1;
+        return false;
     }
 
     /**
@@ -898,8 +830,6 @@ class FormSidebarWidget extends AdminWidget implements
 
         return !!$condition;
     }
-
-
 
     // ACL Permissions
     // =========================================================================
