+++ templates/charcoal/admin/template/form/nav-tabs-languages.mustache
@@ -1,6 +1,6 @@
 {{# showLanguageSwitch }}
 <ul class="c-form-widget_nav nav nav-tabs available-languages" role="tablist">
-    {{# availableLanguages }}
+    {{# languages }}
     <li class="nav-item">
         <a class="nav-link {{# isSelected }} active{{/ isSelected }} js-group-tabs"
            id="{{ widgetId }}_tab_{{ ident }}"
@@ -14,6 +14,6 @@
             <span class="nav-title">{{ name }}</span>
         </a>
     </li>
-    {{/ availableLanguages }}
+    {{/ languages }}
 </ul>
 {{/ showLanguageSwitch }}
